+++ run/CMakeLists.txt
@@ -2,12 +2,12 @@
 cmake_minimum_required (VERSION 3.0.2)
 
 
-# Sets for output directory for executables and libraries. 
+# Sets for output directory for executables and libraries.
 set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/bin)
 set(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/lib)
 set(CMAKE_ARCHIVE_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/lib)
 
-# Sets the position independent code property for all targets. 
+# Sets the position independent code property for all targets.
 set(CMAKE_POSITION_INDEPENDENT_CODE ON)
 
 
@@ -22,7 +22,7 @@ add_definitions(-DWITH_GENX)
 # Creates the EPANET command line executable
 add_executable(runepanet ${EPANET_CLI_SOURCES})
 if(NOT WIN32)
-    target_link_libraries(runepanet LINK_PUBLIC epanet m)
+    target_link_libraries(runepanet LINK_PUBLIC epanet2 m)
 else(NOT WIN32)
-    target_link_libraries(runepanet LINK_PUBLIC epanet)
+    target_link_libraries(runepanet LINK_PUBLIC epanet2)
 endif(NOT WIN32)
